Move edhrec types to types directory
